# Greatly improve database performance for `esgpull update`